Extracting conversion logic from the plugin
